ai-agents/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.1.0
pyyaml>=6.0
requests>=2.31.0
docker>=6.1.0
psycopg2-binary>=2.9.0
neo4j>=5.12.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
openai>=1.0.0
chromadb>=1.1.0
qdrant-client>=1.15.0
elasticsearch>=9.1.0
opensearch-py>=3.0.0
grafana-api>=1.0.0
grafana-dashboard>=0.1.0
grafana-client>=5.0.0
python-dotenv>=1.0.0